requirements.txt
django~=3.1
djangorestframework~=3.11
djangorestframework-simplejwt~=4.4
pyyaml~=5.2
clearbit~=0.1.7
requests~=2.21